requirements.txt
numpy>=1.23
scipy>=1.10
pandas>=1.5
matplotlib>=3.6
scikit-learn>=1.2
torch>=2.0
torchvision>=0.15
torchaudio>=2.0
librosa>=0.10
soundfile>=0.12
opencv-python>=4.7
Pillow>=9.5
albumentations>=1.3
imgaug>=0.4
PyWavelets>=1.4
pyyaml>=6.0
tqdm>=4.65
joblib>=1.3
tensorboard>=2.12
imbalanced-learn>=0.11
seaborn>=0.12
plotly>=5.10
wandb>=0.16
hydra-core>=1.3
numba>=0.57
ruamel.yaml>=0.17
psutil>=5.9
jupyterlab>=4.0
